pyproject.toml
[project]
name = "telescope"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.2",
    "django-cors-headers>=4.5.0",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.3.1",
    "psycopg2-binary>=2.9.10",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
